Preview Tab Sprite Fix and ArtifacTab stuff
widened the previewtab sprite so that there's no weird gap beneath the shadow
More override stuff so that tabmanager isn't such a blob
