+++ Slider/Assets/Scripts/UI/Artifact/Tabs/DesertTabManager.cs
@@ -1,5 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 public class DesertTabManager : ArtifactTabManager
